Fix HOPO calculations in AutoGH
In RockBand games, the time threshold for considering whether a note
is a HOPO can vary. By default, it is 170 ticks but it can be more or
less.  This data is recorded in a config file in the game but is not
stored in the MIDI file.

This change allows the threshold for games and an override for
individual songs to be stored in RhythmGames.xml. The songs are indexed
by title, which is stored as the name of track 0 in RockBand tracks.

If not specified, it continues to use the old formula of a 12th note
(which seems to correspond to 160 for most songs).

This was tested with Black Hole Sun (on Rock Band) and by getting the
100% with HOPO achievement in Beatles Rock Band.

Information on HOPO thresholds came from
https://github.com/ajanta/rock-band-chart-generator

The songs added to the config are only the songs I can validate (the
on-disk Rock Band, Lego Rock Band, and Rock Band Beatles).